给怪物增加追踪功能

+++ Assets/GF_JustOneLevel/Scripts/DataTable/DREntity.cs
@@ -58,7 +58,7 @@ public class DREntity : IDataRow {
     /// 攻击范围
     /// </summary>
     /// <returns></returns>
-    public int AtkRange {
+    public float AtkRange {
         get;
         protected set;
     }
