+++ Assets/GF_JustOneLevel/Scripts/DataTable/DRHero.cs
@@ -15,7 +15,7 @@ public class DRHero : DREntity {
         MoveSpeed = float.Parse (text[index++]);
         RotateSpeed = float.Parse (text[index++]);
         Atk = int.Parse (text[index++]);
-        AtkRange = int.Parse (text[index++]);
+        AtkRange = float.Parse (text[index++]);
         Def = int.Parse (text[index++]);
         AtkSpeed = float.Parse (text[index++]);
         HP = int.Parse (text[index++]);
